Refactoring: Move commands into a `command_variables.ts`
This commit moves all the commands intended to be used as "comand
variables" from the `extension.ts` into a new `command_variables.ts`
file. This is a precursory refactoring before introducing additional
command variables.
